XRDDEV-2187 API on the Central Server to allow processing owner change requests (#1453)
* XRDDEV-2187 setting sourceSets in gradle

* XRDDEV-2187 minor fixes

* XRDDEV-2187 handling owner change request

* XRDDEV-2187 deleting unused code

* XRDDEV-2187 owner change request unit tests

* XRDDEV-2187 service to add/remove global group member

* XRDDEV-2187 adding/removing members from security-server-owners global group after changing the owner

* XRDDEV-2187 minor fixes

* XRDDEV-2187 GlobalGroupMemberService, Identifier Repositories fix

* XRDDEV-2187 Integration test for owner change request.
Variuos issues in other management requests handling fixed.

* XRDDEV-2187 Unused dto classes deleted

* XRDDEV-2187 adding comment to liquibase changeset

+++ src/centralserver/admin-service/core/src/main/java/org/niis/xroad/cs/admin/core/repository/IdentifierRepository.java
@@ -29,7 +29,7 @@ package org.niis.xroad.cs.admin.core.repository;
 import org.niis.xroad.cs.admin.core.entity.XRoadIdEntity;
 
 public interface IdentifierRepository<T extends XRoadIdEntity> extends
-        GenericRepository<T, Long>,
-        FindOrCreateAwareRepository<T, Long> {
+        GenericRepository<T, Integer>,
+        FindOrCreateAwareRepository<T, Integer> {
 
 }
