+++ src/centralserver/admin-service/infra-jpa/src/main/java/org/niis/xroad/cs/admin/jpa/repository/JpaIdentifierRepository.java
@@ -36,8 +36,8 @@ import java.util.List;
 
 @Repository
 public interface JpaIdentifierRepository<T extends XRoadIdEntity> extends
-        JpaRepository<T, Long>,
-        JpaFindOrCreateAwareRepository<T, Long>,
+        JpaRepository<T, Integer>,
+        JpaFindOrCreateAwareRepository<T, Integer>,
         IdentifierRepository<T> {
 
     @Override
